refactor: subdivide utils
